src/main/java/roomescape/repository/rowmapper/ReservationRowMapper.java
package roomescape.repository.rowmapper;

import java.sql.ResultSet;
import java.sql.SQLException;
import org.springframework.jdbc.core.RowMapper;
import org.springframework.stereotype.Component;
import roomescape.entity.Reservation;
import roomescape.entity.ReservationTime;

@Component
public class ReservationRowMapper implements RowMapper<Reservation> {

    @Override
    public Reservation mapRow(ResultSet rs, int rowNum) throws SQLException {
        return new Reservation(
                rs.getLong("reservation_id"),
                rs.getString("name"),
                rs.getString("date"),
                new ReservationTime(
                        rs.getLong("time_id"),
                        rs.getString("time_value")
                )
        );
    }
}
